+++ headers/vector.hpp
@@ -3,15 +3,15 @@
 
 typedef struct
 {
-    double p[2];
+    float p[2];
 } Vec2;
 
-Vec2 Vec2FromScalar(double x)
+Vec2 Vec2FromScalar(float x)
 {
     return (Vec2){{x, x}};
 }
 
-Vec2 Vec2From(double x, double y)
+Vec2 Vec2From(float x, float y)
 {
     return (Vec2){{x, y}};
 }
@@ -43,7 +43,7 @@ Vec2 Vec2Mult(Vec2 v1, Vec2 v2)
     return v1;
 }
 
-Vec2 Vec2MultScalar(Vec2 v, double s)
+Vec2 Vec2MultScalar(Vec2 v, float s)
 {
     for (int i = 0; i < 2; i++)
     {
@@ -52,9 +52,9 @@ Vec2 Vec2MultScalar(Vec2 v, double s)
     return v;
 }
 
-double Vec2Dot(Vec2 v1, Vec2 v2)
+float Vec2Dot(Vec2 v1, Vec2 v2)
 {
-    double s = 0.0;
+    float s = 0.0;
     for (int i = 0; i < 2; i++)
     {
         s += v1.p[i] * v2.p[i];
@@ -62,12 +62,12 @@ double Vec2Dot(Vec2 v1, Vec2 v2)
     return s;
 }
 
-inline double Vec2Length2(const Vec2 v)
+inline float Vec2Length2(const Vec2 v)
 {
     return Vec2Dot(v, v);
 }
 
-inline double Vec2Length(const Vec2 v)
+inline float Vec2Length(const Vec2 v)
 {
     return sqrt(Vec2Dot(v, v));
 }
@@ -81,7 +81,7 @@ inline Vec2 Vec2Pow(Vec2 v1, Vec2 v2)
     return v1;
 }
 
-inline Vec2 Vec2PowS(Vec2 v1, double v2)
+inline Vec2 Vec2PowS(Vec2 v1, float v2)
 {
     for (int i = 0; i < 2; i++)
     {
@@ -99,9 +99,9 @@ inline Vec2 Vec2Max(Vec2 v1, Vec2 v2)
     return v1;
 }
 #ifndef OPENCLCOMP
-inline void Vec2FixUpper(Vec2 *v, double U)
+inline void Vec2FixUpper(Vec2 *v, float U)
 #else
-inline void Vec2FixUpper(global Vec2 *v, double U)
+inline void Vec2FixUpper(global Vec2 *v, float U)
 #endif
 {
     for (int i = 0; i < 2; i++)
@@ -111,9 +111,9 @@ inline void Vec2FixUpper(global Vec2 *v, double U)
 }
 
 #ifndef OPENCLCOMP
-inline void Vec2FixLower(Vec2 *v, double U)
+inline void Vec2FixLower(Vec2 *v, float U)
 #else
-inline void Vec2FixLower(global Vec2 *v, double U)
+inline void Vec2FixLower(global Vec2 *v, float U)
 #endif
 {
     for (int i = 0; i < 2; i++)
@@ -124,15 +124,15 @@ inline void Vec2FixLower(global Vec2 *v, double U)
 
 typedef struct
 {
-    double p[3];
+    float p[3];
 } Vec3;
 
-Vec3 Vec3FromScalar(double x)
+Vec3 Vec3FromScalar(float x)
 {
     return (Vec3){{x, x, x}};
 }
 
-Vec3 Vec3From(double x, double y, double z)
+Vec3 Vec3From(float x, float y, float z)
 {
     return (Vec3){{x, y, z}};
 }
@@ -164,7 +164,7 @@ Vec3 Vec3Mult(Vec3 v1, Vec3 v3)
     return v1;
 }
 
-Vec3 Vec3MultScalar(Vec3 v, double s)
+Vec3 Vec3MultScalar(Vec3 v, float s)
 {
     for (int i = 0; i < 3; i++)
     {
@@ -173,9 +173,9 @@ Vec3 Vec3MultScalar(Vec3 v, double s)
     return v;
 }
 
-double Vec3Dot(Vec3 v1, Vec3 v3)
+float Vec3Dot(Vec3 v1, Vec3 v3)
 {
-    double s = 0.0;
+    float s = 0.0;
     for (int i = 0; i < 3; i++)
     {
         s += v1.p[i] * v3.p[i];
@@ -183,12 +183,12 @@ double Vec3Dot(Vec3 v1, Vec3 v3)
     return s;
 }
 
-inline double Vec3Length2(const Vec3 v)
+inline float Vec3Length2(const Vec3 v)
 {
     return Vec3Dot(v, v);
 }
 
-inline double Vec3Length(const Vec3 v)
+inline float Vec3Length(const Vec3 v)
 {
     return sqrt(Vec3Dot(v, v));
 }
@@ -202,7 +202,7 @@ inline Vec3 Vec3Pow(Vec3 v1, Vec3 v3)
     return v1;
 }
 
-inline Vec3 Vec3PowS(Vec3 v1, double v3)
+inline Vec3 Vec3PowS(Vec3 v1, float v3)
 {
     for (int i = 0; i < 3; i++)
     {
@@ -221,9 +221,9 @@ inline Vec3 Vec3Max(Vec3 v1, Vec3 v3)
 }
 
 #ifndef OPENCLCOMP
-inline void Vec3FixUpper(Vec3 *v, double U)
+inline void Vec3FixUpper(Vec3 *v, float U)
 #else
-inline void Vec3FixUpper(global Vec3 *v, double U)
+inline void Vec3FixUpper(global Vec3 *v, float U)
 #endif
 {
     for (int i = 0; i < 3; i++)
@@ -233,9 +233,9 @@ inline void Vec3FixUpper(global Vec3 *v, double U)
 }
 
 #ifndef OPENCLCOMP
-inline void Vec3FixLower(Vec3 *v, double U)
+inline void Vec3FixLower(Vec3 *v, float U)
 #else
-inline void Vec3FixLower(global Vec3 *v, double U)
+inline void Vec3FixLower(global Vec3 *v, float U)
 #endif
 {
     for (int i = 0; i < 3; i++)
@@ -246,15 +246,15 @@ inline void Vec3FixLower(global Vec3 *v, double U)
 
 typedef struct
 {
-    double p[4];
+    float p[4];
 } Vec4;
 
-Vec4 Vec4FromScalar(double x)
+Vec4 Vec4FromScalar(float x)
 {
     return (Vec4){{x, x, x, x}};
 }
 
-Vec4 Vec4From(double x, double y, double z, double w)
+Vec4 Vec4From(float x, float y, float z, float w)
 {
     return (Vec4){{x, y, z, w}};
 }
@@ -286,7 +286,7 @@ Vec4 Vec4Mult(Vec4 v1, Vec4 v4)
     return v1;
 }
 
-Vec4 Vec4MultScalar(Vec4 v, double s)
+Vec4 Vec4MultScalar(Vec4 v, float s)
 {
     for (int i = 0; i < 4; i++)
     {
@@ -295,9 +295,9 @@ Vec4 Vec4MultScalar(Vec4 v, double s)
     return v;
 }
 
-double Vec4Dot(Vec4 v1, Vec4 v4)
+float Vec4Dot(Vec4 v1, Vec4 v4)
 {
-    double s = 0.0;
+    float s = 0.0;
     for (int i = 0; i < 4; i++)
     {
         s += v1.p[i] * v4.p[i];
@@ -305,12 +305,12 @@ double Vec4Dot(Vec4 v1, Vec4 v4)
     return s;
 }
 
-inline double Vec4Length2(const Vec4 v)
+inline float Vec4Length2(const Vec4 v)
 {
     return Vec4Dot(v, v);
 }
 
-inline double Vec4Length(const Vec4 v)
+inline float Vec4Length(const Vec4 v)
 {
     return sqrt(Vec4Dot(v, v));
 }
@@ -324,7 +324,7 @@ inline Vec4 Vec4Pow(Vec4 v1, Vec4 v4)
     return v1;
 }
 
-inline Vec4 Vec4PowS(Vec4 v1, double v4)
+inline Vec4 Vec4PowS(Vec4 v1, float v4)
 {
     for (int i = 0; i < 4; i++)
     {
@@ -343,9 +343,9 @@ inline Vec4 Vec4Max(Vec4 v1, Vec4 v3)
 }
 
 #ifndef OPENCLCOMP
-inline void Vec4FixUpper(Vec4 *v, double U)
+inline void Vec4FixUpper(Vec4 *v, float U)
 #else
-inline void Vec4FixUpper(global Vec4 *v, double U)
+inline void Vec4FixUpper(global Vec4 *v, float U)
 #endif
 {
     for (int i = 0; i < 4; i++)
@@ -355,9 +355,9 @@ inline void Vec4FixUpper(global Vec4 *v, double U)
 }
 
 #ifndef OPENCLCOMP
-inline void Vec4FixLower(Vec4 *v, double U)
+inline void Vec4FixLower(Vec4 *v, float U)
 #else
-inline void Vec4FixLower(global Vec4 *v, double U)
+inline void Vec4FixLower(global Vec4 *v, float U)
 #endif
 {
     for (int i = 0; i < 4; i++)
